% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz game}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{A positive finite integer.}
}
\value{
A character vector containing the FizzBuzz output from 1 to n.
}
\description{
Vector is created for n terms in the game
If divisible by 3 its replaced with fizz
If divisible by 5 replaced with buzz
If divisible by both then replace with fizz buzz.
}
\examples{
FizzBuzz(15)
}
