% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Flagstaff Pulliam Airport weather data}
\format{
A data frame with 1719 observations and 5 variables:
\describe{
  \item{DATE}{Date of the weather observation.}
  \item{PRCP}{Daily precipitation.}
  \item{SNOW}{Daily snowfall.}
  \item{TMAX}{Daily maximum temperature.}
  \item{TMIN}{Daily minimum temperature.}
}
}
\source{
\url{https://www.ncdc.noaa.gov}
}
\usage{
Flagstaff_Weather
}
\description{
Weather reports each day from 2015 - 2019 from the Flagstaff Pulliam Airport
This is the clean data from Pulliam Airport weather station
}
\keyword{datasets}
